+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.1.14] - 2026-02-14
+## [0.2.0] - 2026-02-16
 
 ### Added
 - `MapFS` implementation
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - `cd()` - returns error now if target path is a file
+- Fixed known bugs
 - Inaccuracies in the documentation have been corrected
 
 ## [0.1.13] - 2026-02-13